Clone cluster alert rule throw an error https://github.com/rancher/rancher/issues/17475
