+++ lib/alert/addon/edit-rule/route.js
@@ -1,10 +1,11 @@
 import Route from '@ember/routing/route';
 import { hash } from 'rsvp';
-import { get, setProperties, set } from '@ember/object';
+import { get } from '@ember/object';
 import { inject as service } from '@ember/service';
 import { reads } from '@ember/object/computed'
+import EditOrClone from 'alert/mixins/edit-or-clone';
 
-export default Route.extend({
+export default Route.extend(EditOrClone, {
   globalStore: service(),
   scope:       service(),
   growl:       service(),
@@ -48,67 +49,6 @@ export default Route.extend({
     }
   },
 
-  getNewClusterAlert(alert) {
-    const gs = get(this, 'globalStore');
-
-    const nodeRule = gs.createRecord({ type: 'nodeRule' });
-    const systemServiceRule = gs.createRecord({ type: 'systemServiceRule' });
-    const eventRule = gs.createRecord({ type: 'eventRule' });
-    const metricRule = gs.createRecord({
-      type:           'metricRule',
-      comparison:     'greater-than',
-      duration:       '5m',
-      thresholdValue: 0,
-    })
-
-    const t = get(alert, 'targetType');
-    const et = get(alert, 'eventRule.eventType');
-
-    set(alert, '_targetType', t);
-
-    switch (t) {
-    case 'event':
-      switch (et) {
-      case 'Normal':
-        set(alert, '_targetType', 'normalEvent');
-        break;
-      case 'Warning':
-        set(alert, '_targetType', 'warningEvent');
-        break;
-      }
-      setProperties(alert, {
-        nodeRule,
-        systemServiceRule,
-        metricRule,
-      });
-      break;
-    case 'node':
-    case 'nodeSelector':
-      setProperties(alert, {
-        eventRule,
-        systemServiceRule,
-        metricRule,
-      });
-      break;
-    case 'systemService':
-      setProperties(alert, {
-        nodeRule,
-        eventRule,
-        metricRule,
-      });
-      break;
-    case 'metric':
-      setProperties(alert, {
-        nodeRule,
-        systemServiceRule,
-        eventRule,
-      })
-      break;
-    }
-
-    return alert;
-  },
-
   loadClusterResource({
     clusterId, ruleId, groupId
   }) {
@@ -139,50 +79,10 @@ export default Route.extend({
         nodes,
         notifiers,
         alertGroup,
-        alertRule: this.getNewClusterAlert(alertRule.clone()),
+        alertRule: this.loadClusterRule(alertRule.clone()),
         metrics:    metrics && metrics.body && metrics.body.names,
       }
     });
-  },
-
-  getNewProjectAlert(alert) {
-    const gs = get(this, 'globalStore');
-    const podRule = gs.createRecord({ type: 'podRule' });
-    const workloadRule = gs.createRecord({ type: 'workloadRule' });
-    const metricRule = gs.createRecord({
-      type:           'metricRule',
-      comparison:     'greater-than',
-      duration:       '5m',
-      thresholdValue:  0,
-    })
-
-    const t = get(alert, 'targetType');
-
-    set(alert, '_targetType', t);
-
-    switch (t) {
-    case 'pod':
-      setProperties(alert, {
-        metricRule,
-        workloadRule,
-      })
-      break;
-    case 'workload':
-    case 'workloadSelector':
-      setProperties(alert, {
-        podRule,
-        metricRule,
-      })
-      break;
-    case 'metric':
-      setProperties(alert, {
-        podRule,
-        workloadRule,
-      })
-      break;
-    }
-
-    return alert;
   },
 
   loadProjectResource({
@@ -223,7 +123,7 @@ export default Route.extend({
         notifiers,
         metrics:    metrics && metrics.body && metrics.body.names,
         alertGroup,
-        alertRule: this.getNewProjectAlert(alertRule.clone()),
+        alertRule: this.loadProjectRule(alertRule.clone()),
       }
     });
   },
