+++ lib/alert/addon/new-rule/route.js
@@ -1,10 +1,11 @@
 import Route from '@ember/routing/route';
 import { hash } from 'rsvp';
-import { get, set, setProperties } from '@ember/object';
+import { get, set } from '@ember/object';
 import { inject as service } from '@ember/service';
 import { reads } from '@ember/object/computed'
+import EditOrClone from 'alert/mixins/edit-or-clone';
 
-export default Route.extend({
+export default Route.extend(EditOrClone, {
   globalStore: service(),
   scope:       service(),
   growl:       service(),
@@ -75,22 +76,8 @@ export default Route.extend({
       newAlert = globalStore.find('clusterAlertRule', id)
         .then( ( alert ) => {
           const cloned =  alert.cloneForNew() ;
-          const t = get(cloned, 'targetType');
 
-          if ( t === 'event' ) {
-            const et = get(cloned, 'eventRule.eventType');
-
-            if ( et === 'Normal' ) {
-              set(cloned, '_targetType', 'normalEvent');
-            }
-            if ( et === 'Warning' ) {
-              set(cloned, '_targetType', 'warningEvent');
-            }
-          } else {
-            set(cloned, '_targetType', t);
-          }
-
-          return cloned;
+          return this.loadClusterRule(cloned);
         });
     } else {
       newAlert = this.getNewClusterAlert(clusterId, groupId);
@@ -154,36 +141,8 @@ export default Route.extend({
       newAlert = globalStore.find('projectAlertRule', id)
         .then( ( alert ) => {
           const cloned = alert.cloneForNew() ;
-          const t = get(cloned, 'targetType');
 
-          set(cloned, '_targetType', t);
-          const workloadRule = globalStore.createRecord({ type: 'workloadRule' })
-          const podRule = globalStore.createRecord({ type: 'podRule' })
-          const metricRule = globalStore.createRecord({ type: 'metricRule' })
-
-          switch (t) {
-          case 'pod':
-            setProperties(cloned, {
-              workloadRule,
-              metricRule,
-            })
-            break;
-          case 'workload':
-          case 'workloadSelector':
-            setProperties(cloned, {
-              podRule,
-              metricRule,
-            })
-            break;
-          case 'metric':
-            setProperties(cloned, {
-              podRule,
-              workloadRule,
-            })
-            break;
-          }
-
-          return cloned;
+          return this.loadProjectRule(cloned);
         });
     } else {
       newAlert = this.getNewProjectAlert(projectId);
